Add files via upload
This is my new logo design.
I used Adobe illustrator to create this design and canva for designing mockups .
My client was a yoga instructor and her requirements were that it has to be meaningful and aesthetic and the fame has to be circular.
So I used moon ,sun and pole star which signifies

Moon : According to the Ashtanga tradition, physical practice ideally changes throughout the lunar cycle to effectively sync the body and the mind with natural rhythms.

Sun: The Sun is called Surya and is considered one of the primary forms of God. Surya is the heart of the world as all other planets revolve around it, it represents willpower, vitality, and courage in your life.

Pole star : It signifies wisdom and longevity science, pointing us toward balance, and a life in perfect alignment with our true character.